pyproject.toml
[tool.poetry]
name = "Stocker"
version = "0.1.0"
description = "Stock Portfolio App"
authors = ["Red Team SE-6"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.5"
django-crispy-forms = "^1.9.0"
matplotlib = "^3.2.1"
pandas = "^1.0.3"
alpha_vantage = "^2.1.3"
djangorestframework = "^3.11.0"
ipdb = "^0.13.2"
django-cors-headers = "^3.2.1"

[tool.poetry.dev-dependencies]
poetry = "^1.0.5"
djangorestframework = "^3.11.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
